sample-python/requirements.txt
streamlit==1.36.0
pyyaml==6.0.1
PyGithub==2.3.0
requests==2.32.3
python-dotenv==1.0.1
